+++ tests/test_structure_data.py
@@ -83,6 +83,21 @@ class TestStructureData(unittest.TestCase):
             split_sites_data = StructureData.split_sites_data(sites_data=mock_sites_data,
                 x_limits=x_limits)
 
+    # @patch('pyscses.structure_data.sites_data_from_file')
+    # def test_StructureData_from_file(self,
+    #                                  mock_sites_data_from_file):
+    #     filename = 'foo'
+    #     x_limits = (0.0, 1.0)
+    #     b = 1.0
+    #     c = 1.0
+    #     system = 'single'
+    #     StructureData.from_file(filename=filename,
+    #                             x_limits=x_limits,
+    #                             b=b,
+    #                             c=c,
+    #                             system=system)
+
+
 
 if __name__ == '__main__':
     unittest.main()
